+++ quasar.config.js
@@ -33,13 +33,13 @@ module.exports = configure(function (/* ctx */) {
       // 'ionicons-v4',
       // 'mdi-v7',
       // 'fontawesome-v6',
-      'eva-icons'
+      // 'eva-icons'
       // 'themify',
       // 'line-awesome',
       // 'roboto-font-latin-ext', // this or either 'roboto-font', NEVER both!
 
       // 'roboto-font' // optional, you are not bound to it
-      // 'material-icons' // optional, you are not bound to it
+      'material-icons' // optional, you are not bound to it
     ],
 
     // Full list of options: https://v2.quasar.dev/quasar-cli-vite/quasar-config-js#build
