+++ src/router/routes.js
@@ -5,8 +5,15 @@ const routes = [
     children: [
 
       { path: '', redirect: '/dashboard' },
-      { path: 'dashboard', component: () => import('pages/Dashboard.vue') },
-      { path: 'profiles', component: () => import('pages/ProfilesPage.vue') },
+      { path: 'dashboard', component: () => import('pages/DashboardPage.vue') },
+      {
+        path: 'profiles',
+        component: () => import('pages/ProfilesPage.vue'),
+        children: [
+          { path: '', component: () => import('pages/profiles/IndexPage.vue') },
+          { path: 'add', component: () => import('pages/profiles/AddProfilePage.vue') }
+        ]
+      },
       { path: 'calculator', component: () => import('pages/CalculatorPage.vue') },
       { path: 'settings', component: () => import('pages/SettingsPage.vue') },
       { path: 'about', component: () => import('pages/AboutPage.vue') }
